Finished calculations for getting the position for the ground below.
Items will spread out, not fall out of the map, go up on platforms if the platform above is closer than the platform below, and fall below if no platform to catch it. I calculate VRLeft and VRRight when the map is loaded since most of the time it was 0 for both. This value is used so items never fall out of the map.